+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 ### :house: Internal
 - `viewer`
   - [b7177a60](https://github.com/wix-incubator/rich-content/commit/b7177a60) text-utils refactoring
+- `example`
+  - [#314](https://github.com/wix-incubator/rich-content/pull/314) content state data validation improved
 <hr/>
 
 ## 3.5.1 (Jun 10, 2019)